@@ -5,7 +5,7 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>1.0.0.RC3</version>
+        <version>1.0.1.RELEASE</version>
     </parent>
 
     <groupId>org.demo</groupId>
@@ -124,7 +124,7 @@
 
     <properties>
         <start-class>demo.Application</start-class>
-        <java.version>1.7</java.version>
+        <java.version>1.8</java.version>
     </properties>
 
     <repositories>
